refact: drag and drop hook
